Moved bundle related code to their own file.
